refactor(tui): extract input overlay runtime component
